fix app page scroll bug renamed folders
