+++ ImPresent.Web/ClientApp/src/app/shared/components/dialogs/loading/loading-dialog.component.scss
