+++ ImPresent.Web/ClientApp/src/app/shared/components/page/page.component.html
@@ -1,4 +1,4 @@
-<div *ngIf="loaded" class="w-full bg-white p-4 rounded-lg select-none max-w-xs md:max-w-2xl break-words overflow-y-auto" style="max-height: 80vh;">
+<div *ngIf="loaded" class="w-full bg-white p-4 rounded-lg select-none max-w-xs md:max-w-2xl break-words" style="max-height: 80vh;">
   <div *ngIf="error == null && success == null">
     <ng-content></ng-content>
   </div>
