+++ ImPresent.Web/ClientApp/src/app/shared/shared.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 
-import { InputDirective } from './directives/input-directive/input.directive';
-import { ButtonDirective } from './directives/button-directive/button.directive';
-import { LoadingDialogComponent } from './components/dialogs/loading-dialog/loading-dialog.component';
+import { InputDirective } from './directives/input/input.directive';
+import { ButtonDirective } from './directives/button/button.directive';
+import { LoadingDialogComponent } from './components/dialogs/loading/loading-dialog.component';
 import { PageComponent } from './components/page/page.component';
 
 @NgModule({
